+++ extensions/notion/src/create-database-page.tsx
@@ -1,6 +1,3 @@
-import { withAccessToken } from "@raycast/utils";
-
 import { CreatePageForm } from "./components/forms/CreatePageForm";
-import { notionService } from "./utils/notion/oauth";
 
-export default withAccessToken(notionService)(CreatePageForm);
+export default CreatePageForm;